requirements.txt
gunicorn~=20.1.0
https://download.pytorch.org/whl/cpu/torch-1.8.1%2Bcpu-cp39-cp39-linux_x86_64.whl
Pillow~=8.2.0
Flask~=2.0.0
httplib2~=0.19.1
oauth2client~=4.1.3
google-api-python-client~=2.5.0

#not clarified but used
opencv-python-headless~=4.5.2.52
pyzbar~=0.1.8
PyYAML~=5.4.1
matplotlib~=3.4.2
pandas~=1.2.4
seaborn~=0.11.1
tensorboard~=2.4.1
PyMySQL~=1.0.2